requirements.txt
amqp==5.2.0
annotated-types==0.6.0
APScheduler==3.10.4
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.4
django-apscheduler==0.6.2
djangorestframework==3.15.1
dnspython==2.6.1
email_validator==2.1.1
graphql_query==1.3.2
idna==3.7
Jinja2==3.1.4
kombu==5.3.7
MarkupSafe==2.1.5
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.5
requests==2.32.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
